(feat):continuacion del proyecto

+++ 14_Proyecto_para_evaluacion_profe/src/modelo/dao/EmpleadoDao.java
@@ -1,6 +1,6 @@
 package modelo.dao;
 
-import java.math.BigDecimal;
+
 import java.util.List;
 
 import modelo.entidades.Empleado;
@@ -11,8 +11,8 @@ public interface EmpleadoDao extends InterfacegenericoCrud<String,Empleado> {
 	 List<Empleado>empleadosByDepartamento(int idDepar);
 	 List<Empleado>empleadosBySexo(char sexo);
 	 List<Empleado>empleadosByApellido(String subcadena);
-	 BigDecimal salarioTotal();
-	 BigDecimal salarioTotal(int idDepar);
+	 double salarioTotal();
+	 double salarioTotal(int idDepar);
 	 
 	
 }
